Update arslatest.php
Though I risk to meet hysterics again, I will try.

I tried to use the plugin to generate several buttons. When the joomla session is clear it works perfect.
I use code like this:
<p><a class="btn btn-danger" href="{arslatest item_link '*.zip' Category 1}">Завантажити останню версію</a></p>
<p><a class="btn btn-danger" href="{arslatest item_link '*.zip' Category 2}">Завантажити останню версію</a></p>

If I browse ars and later return to the page with the buttons, only one buttons works - other links are empty. This is because when I browse ars, the category I browse is stored to joomla session or something like that. So when I return to the page with several buttons, the model filters use the latest category I was browsing.

By setting ->category_id(0) I make the model to look throughout all categories again. Maybe this change kills some caching idea. But I cannot find any other way to make the plugin work with several buttons.