Convert view templates to Blade
Use Chosen where it makes sense in the backend

That's different than the original implementation that used Chosen
everywhere and made things unnecessarily complicated. Yeah, this
solution is not visually consistent but it's really darn usable. The
only reason we get in the backend of this component is to get stuff
done, fast. So there.