# [HIGH] Joomla 4 does not ignore the menu item query when calling parse() in the SEF router
Closes gh-178

Our router worked perfectly. The problem is that Joomla 4
merges the query array of the menu item with the `$query`
array we return from the SEF router's `parse()` method.
If you have a Category menu item you have the `layout`
of the Category item being applied (normal, blledingedge or
respoitory). However, this layout does not exist in the
Releases and Items views, causing an internal error.

Joomla 3 was not susceptible to this because the query
we returned overrode the one from the menu.

This issue was complicated to pin down but it has a simple,
one liner fix. Just set the layout to default...